inference/requirements.txt
torch
torchvision
deeplake
Flask
flask-cors
numpy
pillow